requirements.txt
psycopg2-binary==2.8.6
python-dotenv==0.15.0
termcolor==1.1.0